Properly implement cas methods, fix some of the broken tests.
